// packages/contracts/src/generated/exchange.ts
export const SuppliersRow = z.object({
  "id": z.string().uuid(),
  "name": z.string(),
  "email": z.string().nullable(),
  "phone": z.string().nullable(),
  "created_at": z.string().nullable(),
  "updated_at": z.string().nullable(),
  "logo": z.string().nullable(),
  "is_active": z.boolean().nullable(),
});
export type SuppliersRow = z.infer<typeof SuppliersRow>;